relo
relocate user table page

add title dashboards/tables
