+++ pages/dashboard/tables.tsx
@@ -1,9 +1,14 @@
 import MainLayout from '@common/components/Layouts/Main'
 import DashboardPage from '@common/components/DashboardPage'
 import { AppRoutes } from '@utils/constants'
+import Head from 'next/head'
 
 export default function TablesPage() {
   return (
+    <>
+    <Head>
+      <title>Всі таблиці</title>
+    </Head>
     <MainLayout
       path={[
         { title: 'Панель управління', path: AppRoutes.INDEX },
@@ -12,5 +17,6 @@ export default function TablesPage() {
     >
       <DashboardPage />
     </MainLayout>
+    </>
   )
 }
